requirements.txt
streamlit==1.46.1
pandas
matplotlib
pillow
numpy
opencv-python-headless
mediapipe
openai
python-dotenv
pytest
